+++ packages/core/core/src/requests/DevDepRequest.js
@@ -108,7 +108,7 @@ export async function getDevDepRequests<TResult>(
     await Promise.all(
       api
         .getSubRequests()
-        .filter(req => req.type === 'dev_dep_request')
+        .filter(req => req.requestType === 'dev_dep_request')
         .map(async req => [
           req.id,
           nullthrows(await api.getRequestResult<DevDepRequest>(req.id)),