+++ packages/packagers/js/src/helpers.js
@@ -112,7 +112,7 @@ function $parcel$export(e, n, v, s) {
 const $parcel$exportWildcard = `
 function $parcel$exportWildcard(dest, source) {
   Object.keys(source).forEach(function(key) {
-    if (key === 'default' || key === '__esModule' || dest.hasOwnProperty(key)) {
+    if (key === 'default' || key === '__esModule' || Object.prototype.hasOwnProperty.call(dest, key)) {
       return;
     }