+++ packages/reporters/cli/src/CLIReporter.js
@@ -15,6 +15,7 @@ import chalk from 'chalk';
 import {getTerminalWidth} from './utils';
 import logLevels from './logLevels';
 import bundleReport from './bundleReport';
+import phaseReport from './phaseReport';
 import {
   writeOut,
   updateSpinner,
@@ -30,6 +31,8 @@ import cliProgress from 'cli-progress';
 const THROTTLE_DELAY = 100;
 const seenWarnings = new Set();
 const seenPhases = new Set();
+const seenPhasesGen = new Set();
+const phaseStartTimes = {};
 
 let statusThrottle = throttle((message: string) => {
   updateSpinner(message);
@@ -73,6 +76,11 @@ export async function _report(
         break;
       }
 
+      if (!seenPhasesGen.has(event.phase)) {
+        phaseStartTimes[event.phase] = Date.now();
+        seenPhasesGen.add(event.phase);
+      }
+
       if (!isTTY && logLevelFilter != logLevels.verbose) {
         if (event.phase == 'transforming' && !seenPhases.has('transforming')) {
           updateSpinner('Building...');
@@ -86,7 +94,6 @@ export async function _report(
           updateSpinner('Packaging & Optimizing...');
         }
         seenPhases.add(event.phase);
-
         break;
       }
 
@@ -105,6 +112,8 @@ export async function _report(
         break;
       }
 
+      phaseStartTimes['buildSuccess'] = Date.now();
+
       persistSpinner(
         'buildProgress',
         'success',
@@ -118,6 +127,9 @@ export async function _report(
           options.projectRoot,
           options.detailedReport?.assetsPerBundle,
         );
+        if (process.env.PARCEL_SHOW_PHASE_TIMES) {
+          phaseReport(phaseStartTimes);
+        }
       }
       break;
     case 'buildFailure':